requirements.txt
networkx==3.6
langchain==0.3.13
langchain-community==0.3.13
matplotlib==3.9.3
python-dotenv==1.0.1
pytest==8.3.4
